+++ locales/ar.js
@@ -1,14 +1,14 @@
 /* ar - ARABIC Translation by RADHWANEDZ
  Translation strings used by the bot
- Strings wrapped in {{ }} will be auto-replaced with updated values
- Globals: {{servername}}, {{invite}}, (all from config) and {{playercount}} (current connected players)
+ Strings wrapped in { } will be auto-replaced with updated values
+ Globals: {servername}, {invite}, (all from config) and {playercount} (current connected players)
 */
 const locale = {
-    checkingWhitelist: "يتم الان التحقق من حسابك اذا كان مفعل ام لا ،{{name}} مرحبا",
+    checkingWhitelist: "يتم الان التحقق من حسابك اذا كان مفعل ام لا ،{name} مرحبا",
     discordNotOpen: "يجب ان تقوم بفتح تطبيق الديسكورد قبل فتح برنامج فايف ام ",
     fatalError: "نعتقد ان هناك خطأ ما في التحقق من تفعيل حسابك يرجى مراسلتنا عبر الديسكورد لحل المشكلة",
-    notInDiscordServer: " {{invite}} يرجى الانضمام الي الديسكورد وتفعيل عضويتك، اليك الرابط {{servername}} انت لست عضوا في سيرفر الديسكورد الخاص بنا ",
-    notWhitelisted: "ليتم تفعيل حسابك لتستطيع الدخول للمدينة {{servername}} يجب اجتياز المقابلة في الديسكورد اولا ",
+    notInDiscordServer: " {invite} يرجى الانضمام الي الديسكورد وتفعيل عضويتك، اليك الرابط {servername} انت لست عضوا في سيرفر الديسكورد الخاص بنا ",
+    notWhitelisted: "ليتم تفعيل حسابك لتستطيع الدخول للمدينة {servername} يجب اجتياز المقابلة في الديسكورد اولا ",
     kickedWithoutReason: "لقد تم طردك من قبل المراقب.",
     announcement: "اعلان",
     directMessage: "الادارة",