+++ locales/de.js
@@ -1,14 +1,14 @@
 /* de - German Translation by Anonymous
  Translation strings used by the bot
- Strings wrapped in {{ }} will be auto-replaced with updated values
-Globals: {{servername}}, {{invite}}, (all from config) and {{playercount}} (current connected players)
+ Strings wrapped in { } will be auto-replaced with updated values
+Globals: {servername}, {invite}, (all from config) and {playercount} (current connected players)
 */
 const locale = {
-    checkingWhitelist: "Hi {{name}}, wir überprüfen deinen Whitelist status..",
+    checkingWhitelist: "Hi {name}, wir überprüfen deinen Whitelist status..",
     discordNotOpen: "Du musst Discord offen haben, bevor du FiveM startest..(Starte beide neu wenn ein Problem auftritt)",
     fatalError: "Beim Abrufen Ihres Whitelist-Status ist ein Fehler aufgetreten. Bitte versuchen Sie es in Kürze erneut oder kontaktieren Sie den Support im Discord, wenn das Problem weiterhin besteht.",
-    notInDiscordServer: "Du bist nicht in dem {{servername}} Discord, tritt hier bei: {{invite}}",
-    notWhitelisted: "Du hast nicht die {{servername}} Discord-Rolle, die benötigt wird um dem Server beizutreten, bist du auf der Whitelist?",
+    notInDiscordServer: "Du bist nicht in dem {servername} Discord, tritt hier bei: {invite}",
+    notWhitelisted: "Du hast nicht die {servername} Discord-Rolle, die benötigt wird um dem Server beizutreten, bist du auf der Whitelist?",
     kickedWithoutReason: "Du wurdest vom Team gekickt.",
     announcement: "ANNOUNCEMENT",
     directMessage: "STAFF",